Revert "Make emit_metrics importable without having boto3 installed"
This reverts commit 86109ab54a4204a2c20c35e35c9c83d114e18c5f.

[ghstack-poisoned]
